+++ src/s_tir/meta_schedule/database/json_database.cc
@@ -195,7 +195,7 @@ Database Database::JSONDatabase(ffi::String path_workload, ffi::String path_tuni
           try {
             const ffi::ArrayObj* arr = json_obj.as<ffi::ArrayObj>();
             TVM_FFI_ICHECK_EQ(arr->size(), 2);
-            int64_t workload_index = arr->at(0).cast<IntImm>()->value;
+            int64_t workload_index = static_cast<int64_t>(arr->at(0).cast<IntImm>()->value);
             TVM_FFI_ICHECK(workload_index >= 0 &&
                            static_cast<size_t>(workload_index) < workloads.size());
             workload = workloads[workload_index];
