+++ src/s_tir/transform/default_gpu_schedule.cc
@@ -66,7 +66,7 @@ void ThreadBind(s_tir::Schedule sch, const s_tir::SBlockRV& block, int64_t max_t
   s_tir::LoopRV fused = sch->Fuse(data_parallel_loops, /*preserve_unit_iters=*/false);
   int64_t product = std::numeric_limits<int64_t>::max();
   if (sch->Get(fused)->extent->IsInstance<IntImmNode>()) {
-    product = sch->Get(fused)->extent.as<IntImmNode>()->value;
+    product = static_cast<int64_t>(sch->Get(fused)->extent.as<IntImmNode>()->value);
   }
   // schedule the fused loop
   if (product > max_thread_per_block * max_threadblocks) {
