+++ src/tirx/transform/narrow_datatype.cc
@@ -92,8 +92,10 @@ class DataTypeVisitor final : public StmtExprVisitor {
         analyzer_->const_int_bound(e, &bound_);
       }
       ConstIntBound bound = bound_[e];
-      int64_t ubound = max_value(PrimType::Int(target_bits_)).as_or_throw<IntImm>()->value;
-      int64_t lbound = min_value(PrimType::Int(target_bits_)).as_or_throw<IntImm>()->value;
+      int64_t ubound =
+          static_cast<int64_t>(max_value(PrimType::Int(target_bits_)).as_or_throw<IntImm>()->value);
+      int64_t lbound =
+          static_cast<int64_t>(min_value(PrimType::Int(target_bits_)).as_or_throw<IntImm>()->value);
       if (e_ty.bits() <= target_bits_ ||
           (bound->max_value <= ubound && bound->min_value >= lbound)) {
         bits = target_bits_;
